Using services with Spring Data Jpa realization

+++ src/main/java/ivan/vatlin/services/CarBaseService.java
@@ -39,7 +39,7 @@ public class CarBaseService implements CarService{
     }
 
     @Override
-    public int getNumberOfCars() {
+    public long getNumberOfCars() {
         return carDao.getNumberOfCars();
     }
 }
